Remove moment dependency from guessTimezone
